requirements.txt
yfinance
requests
pandas
python-dotenv